Move normalize_value to utils
